+++ ProjectBarnabus/src/GameEngine/Camera.h
@@ -13,13 +13,13 @@ class Camera : public Component
 {
 protected:
 	// The camera's forward-facing direction
-	glm::dvec3 forward;
+	glm::vec3 forward;
 	// The camera's current orientation
-	glm::dvec3 up;
+	glm::vec3 up;
 	// The vector perpendicular to both the forward-facing direction and the orientation
-	glm::dvec3 right;
+	glm::vec3 right;
 	// The camera's current target
-	glm::dvec3 target;
+	glm::vec3 target;
 	// The currently built view matrix since the last frame update
 	glm::mat4 view;
 	// The currently built projection matrix since the last call to set_projection
@@ -28,42 +28,42 @@ protected:
 public:
 	// Create a new camera (called by sub-classes)
 	Camera(std::string camType) : 
-		target(glm::dvec3(0.0f, 0.0f, -1.0f)), 
-		up(glm::dvec3(0.0f, 1.0f, 0.0f)),
+		target(glm::vec3(0.0f, 0.0f, -1.0f)), 
+		up(glm::vec3(0.0f, 1.0f, 0.0f)),
 		view(), projection(), Component(camType) {};
 
 	// Destroy the camera (virtual because this is an abstract class)
 	virtual ~Camera() {}
 
-	const glm::dvec3& GetForward() const
+	const glm::vec3& GetForward() const
 	{
 		return forward;
 	}
 
 	// Get camera's orientation
-	const glm::dvec3& GetOrientation() const
+	const glm::vec3& GetOrientation() const
 	{
 		return up;
 	}
 	// Set camera's orientation
-	void SetOrientation(const glm::dvec3 &value)
+	void SetOrientation(const glm::vec3 &value)
 	{
 		up = value;
 	}
 
-	const glm::dvec3& GetRight() const 
+	const glm::vec3& GetRight() const 
 	{ 
 		return right; 
 	}
 
 	// Get camera's current target
-	const glm::dvec3& GetTarget() const
+	const glm::vec3& GetTarget() const
 	{
 		return target;
 	}
 
 	// Set camera's target
-	void SetTarget(const glm::dvec3 &value)
+	void SetTarget(const glm::vec3 &value)
 	{
 		target = value;
 	}
@@ -80,5 +80,5 @@ public:
 	}
 
 	// Updates the camera (pure virtual function)
-	 virtual void Update(double deltaTime) override = 0 ;
+	 virtual void Update(float deltaTime) override = 0 ;
 };
